renamed app to server to prepare for frontend, changed relative routes in server.js, fixed missing script due to faulty merge
