+++ package.json
@@ -4,7 +4,8 @@
   "description": "A beer rating application",
   "main": "server.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "dev": "mongod | nodemon ./server/server.js"
   },
   "author": "Ryan Will",
   "license": "MIT",
